requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.6
aiofiles==23.2.1
openai==2.6.1
python-dotenv==1.0.0
pydantic==2.12.3
pydantic_core==2.41.4
typing_extensions==4.15.0
